requirements.txt
asgiref==3.4.1
cachetools==4.2.2
certifi==2021.5.30
chardet==4.0.0
Django==3.2.5
django-cors-headers==3.7.0
google-auth==1.32.1
gunicorn==20.1.0
idna==2.10
j2cli==0.3.10
Jinja2==3.0.1
kubernetes==17.17.0
MarkupSafe==2.0.1
oauthlib==3.1.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
python-dateutil==2.8.1
python-dotenv==0.19.0
pytz==2021.1
PyYAML==5.4.1
requests==2.25.1
requests-oauthlib==1.3.0
rsa==4.7.2
six==1.16.0
sqlparse==0.4.1
urllib3==1.26.6
websocket-client==1.1.0
whitenoise==5.3.0
